Revert "Refactor Streamlit app into legacy package and scaffold Django backend"
