+++ app/pages/4_History_Reports.py
@@ -1,4 +1,4 @@
-# legacy_streamlit/pages/4_History_Reports.py
+# app/pages/4_History_Reports.py
 import os
 import sys
 import streamlit as st
@@ -13,10 +13,10 @@ if _REPO_ROOT not in sys.path:
     sys.path.insert(0, _REPO_ROOT)
 
 try:
-    from legacy_streamlit.db.database_utils import connect_db
-    from legacy_streamlit.services import item_service
-    from legacy_streamlit.services import stock_service
-    from legacy_streamlit.core.constants import (
+    from app.db.database_utils import connect_db
+    from app.services import item_service
+    from app.services import stock_service
+    from app.core.constants import (
         TX_RECEIVING,
         TX_ADJUSTMENT,
         TX_WASTAGE,
@@ -25,9 +25,9 @@ try:
         PLACEHOLDER_SELECT_ITEM,
         FILTER_ALL_TYPES,
     )
-    from legacy_streamlit.ui.theme import load_css, render_sidebar_logo
-    from legacy_streamlit.ui.navigation import render_sidebar_nav
-    from legacy_streamlit.ui import show_success, show_error, show_warning
+    from app.ui.theme import load_css, render_sidebar_logo
+    from app.ui.navigation import render_sidebar_nav
+    from app.ui import show_success, show_error, show_warning
 except ImportError as e:
     show_error(f"Import error in 4_History_Reports.py: {e}.")
     st.stop()
