+++ app/services/purchase_order_service.py
@@ -1,4 +1,4 @@
-# legacy_streamlit/services/purchase_order_service.py
+# app/services/purchase_order_service.py
 
 import traceback
 from typing import Optional, Dict, List, Tuple, Any
@@ -9,9 +9,9 @@ from sqlalchemy import text
 from sqlalchemy.exc import IntegrityError, SQLAlchemyError
 from sqlalchemy.engine import Engine
 
-from legacy_streamlit.core.logging import get_logger
-from legacy_streamlit.db.database_utils import fetch_data
-from legacy_streamlit.core.constants import (
+from app.core.logging import get_logger
+from app.db.database_utils import fetch_data
+from app.core.constants import (
     PO_STATUS_DRAFT,
     PO_STATUS_ORDERED,
     PO_STATUS_FULLY_RECEIVED,
