+++ app/services/recipe_service.py
@@ -7,9 +7,9 @@ from sqlalchemy import text
 from sqlalchemy.exc import IntegrityError, SQLAlchemyError
 from sqlalchemy.engine import Engine, Connection
 
-from legacy_streamlit.core.logging import get_logger
-from legacy_streamlit.db.database_utils import fetch_data
-from legacy_streamlit.core.constants import TX_SALE
+from app.core.logging import get_logger
+from app.db.database_utils import fetch_data
+from app.core.constants import TX_SALE
 from . import stock_service
 
 logger = get_logger(__name__)
