+++ app/ui/navigation.py
@@ -1,11 +1,11 @@
 import streamlit as st
-from legacy_streamlit.core.logging import flush_logs
+from app.core.logging import flush_logs
 
 
 def render_sidebar_nav(include_clear_logs_button: bool = True) -> None:
     """Render sidebar navigation links to all app pages."""
     with st.sidebar:
-        st.page_link("item_manager_legacy_streamlit.py", label="🏠 Dashboard")
+        st.page_link("item_manager_app.py", label="🏠 Dashboard")
         st.page_link("pages/1_Items.py", label="Items")
         st.page_link("pages/2_Suppliers.py", label="Suppliers")
         st.page_link("pages/3_Stock_Movements.py", label="Stock Movements")
