--- backend/inventory/__init__.py
@@ -1,1 +1,0 @@
-# Django project package
