+++ requirements.txt
@@ -5,7 +5,3 @@ psycopg2-binary # Or psycopg2 if not using binary
 fpdf2
 pytest
 pyyaml
-Django
-djangorestframework
-django-htmx
-django-tailwind
