+++ tests/conftest.py
@@ -9,7 +9,7 @@ PROJECT_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 if PROJECT_ROOT not in sys.path:
     sys.path.insert(0, PROJECT_ROOT)
 
-from legacy_streamlit.core.logging import configure_logging
+from app.core.logging import configure_logging
 
 # Configure logging for tests once before other modules import loggers
 configure_logging()
